pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "unified-agent-framework"
version = "5.0.0"
description = "Runner-Centric AI Agent Framework — run_agent() 한 줄로 끝. Top 3 엔진(Direct, LangChain, CrewAI) + OTEL 표준 모니터링."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Unified Agent Framework Contributors"}
]
keywords = [
    "ai",
    "agent",
    "llm",
    "openai",
    "azure",
    "langchain",
    "crewai",
    "multi-agent",
    "orchestration",
    "runner",
    "mcp",
    "opentelemetry",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
langchain = [
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-core>=0.2.0",
]
crewai = [
    "crewai>=0.50.0",
]
otel = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
]
full = [
    "unified-agent-framework[langchain,crewai,otel]",
]

[project.urls]
Homepage = "https://github.com/unified-agent-framework/unified-agent-framework"
Documentation = "https://github.com/unified-agent-framework/unified-agent-framework#readme"
Repository = "https://github.com/unified-agent-framework/unified-agent-framework.git"
Issues = "https://github.com/unified-agent-framework/unified-agent-framework/issues"
Changelog = "https://github.com/unified-agent-framework/unified-agent-framework/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
where = ["."]
include = ["unified_agent_v5*"]

[tool.setuptools.package-data]
unified_agent_v5 = ["py.typed"]

# ─────────────────────────────────────────────────────────────────────────────────
# Black (Code Formatter)
# ─────────────────────────────────────────────────────────────────────────────────
[tool.black]
line-length = 100
target-version = ['py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
    | \.venv
    | build
    | dist
    | __pycache__
)/
'''

# ─────────────────────────────────────────────────────────────────────────────────
# isort (Import Sorter)
# ─────────────────────────────────────────────────────────────────────────────────
[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["unified_agent_v5"]
skip = [".venv", "build", "dist"]

# ─────────────────────────────────────────────────────────────────────────────────
# Ruff (Linter)
# ─────────────────────────────────────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py

# ─────────────────────────────────────────────────────────────────────────────────
# mypy (Type Checker)
# ─────────────────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# ─────────────────────────────────────────────────────────────────────────────────
# pytest
# ─────────────────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests", "."]
python_files = ["test_*.py", "demo_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

# ─────────────────────────────────────────────────────────────────────────────────
# coverage
# ─────────────────────────────────────────────────────────────────────────────────
[tool.coverage.run]
source = ["unified_agent_v5"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
